requirements.txt
arrow==1.2.3
attrs==22.1.0
boto3==1.26.1
botocore==1.29.1
bravado==11.0.3
bravado-core==5.17.1
catboost==1.1.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.2
fonttools==4.38.0
fqdn==1.5.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
graphviz==0.20.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
isoduration==20.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.3
jsonref==1.0.0.post1
jsonschema==4.17.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
monotonic==1.6
msgpack==1.0.4
neptune-client==0.16.11
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
plotly==5.11.0
psutil==5.9.3
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
s3transfer==0.6.0
scipy==1.9.3
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
swagger-spec-validator==3.0.3
tenacity==8.1.0
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.12
webcolors==1.12
websocket-client==1.4.1
Werkzeug==2.2.2
zipp==3.10.0
